Adds file for BC and SS conditions
